pre branch

+++ src/types/cliente.rs
@@ -1,9 +1,9 @@
-use rocket::serde::Deserialize;
+use rocket::serde::{Deserialize, Serialize};
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize, Serialize, Debug)]
 #[serde(crate = "rocket::serde")]
 pub struct Cliente {
-	pub tipo_doc:i8,
-	pub num_doc:i32,
+	pub tipo_doc:i32,
+	pub num_doc:i64,
 	pub nombre:String
 }
